@lgbo-ustc 2022-03-04
the tests failed are
```
test_system_logs_recreate/test.py::test_system_logs_recreate
```

Not releated to this pr ?