@kssenii 2022-03-08
@lgbo-ustc 
```
2022-03-08 07:45:44,646 runner output: ___________________ ERROR collecting test_hive_query/test.py ___________________
2022-03-08 07:45:44,646 runner output: /usr/local/lib/python3.8/dist-packages/_pytest/python.py:599: in _importtestmodule
2022-03-08 07:45:44,646 runner output:     mod = import_path(self.path, mode=importmode, root=self.config.rootpath)
2022-03-08 07:45:44,646 runner output: /usr/local/lib/python3.8/dist-packages/_pytest/pathlib.py:533: in import_path
2022-03-08 07:45:44,646 runner output:     importlib.import_module(module_name)
2022-03-08 07:45:44,646 runner output: /usr/lib/python3.8/importlib/__init__.py:127: in import_module
2022-03-08 07:45:44,646 runner output:     return _bootstrap._gcd_import(name[level:], package, level)
2022-03-08 07:45:44,646 runner output: <frozen importlib._bootstrap>:1014: in _gcd_import
2022-03-08 07:45:44,646 runner output:     ???
2022-03-08 07:45:44,646 runner output: <frozen importlib._bootstrap>:991: in _find_and_load
2022-03-08 07:45:44,646 runner output:     ???
2022-03-08 07:45:44,646 runner output: <frozen importlib._bootstrap>:975: in _find_and_load_unlocked
2022-03-08 07:45:44,646 runner output:     ???
2022-03-08 07:45:44,646 runner output: <frozen importlib._bootstrap>:671: in _load_unlocked
2022-03-08 07:45:44,646 runner output:     ???
2022-03-08 07:45:44,646 runner output: /usr/local/lib/python3.8/dist-packages/_pytest/assertion/rewrite.py:162: in exec_module
2022-03-08 07:45:44,646 runner output:     source_stat, co = _rewrite_test(fn, self.config)
2022-03-08 07:45:44,646 runner output: /usr/local/lib/python3.8/dist-packages/_pytest/assertion/rewrite.py:364: in _rewrite_test
2022-03-08 07:45:44,646 runner output:     tree = ast.parse(source, filename=strfn)
2022-03-08 07:45:44,646 runner output: /usr/lib/python3.8/ast.py:47: in parse
2022-03-08 07:45:44,646 runner output:     return compile(source, filename, mode, flags,
2022-03-08 07:45:44,646 runner output: E     File "/ClickHouse/tests/integration/test_hive_query/test.py", line 52
2022-03-08 07:45:44,646 runner output: E       logging.info("create result {}".format(result))
2022-03-08 07:45:44,646 runner output: E       ^
2022-03-08 07:45:44,646 runner output: E   IndentationError: unexpected indent
```