@kssenii 2022-03-08
> test_hive_query/test.py::test_cache_read_bytes

Fails

@lgbo-ustc