@lgbo-ustc 2022-03-09
It's strange. `system.metric_log` doesn't exists
```
E           helpers.client.QueryRuntimeException: Client failed! Return code: 60, stderr: Received exception from server (version 22.3.1):
E           Code: 60. DB::Exception: Received from 172.16.1.3:9000. DB::Exception: Table system.metric_log doesn't exist. Stack trace:
```