@kssenii 2022-03-09
> It's strange. system.metric_log doesn't exist

need to add query in test: `system flush logs` (instead of retries)